refactor(shared-docs): use utils pattern for cookie consent function
