+++ docs/components/cookie-popup/cookie-popup.component.ts
@@ -9,6 +9,7 @@
 import {ChangeDetectionStrategy, Component, inject, signal} from '@angular/core';
 import {NgIf} from '@angular/common';
 import {LOCAL_STORAGE} from '../../providers/index';
+import {setCookieConsent} from '../../utils/analytics.utils';
 
 /**
  * Decelare gtag as part of the window in this file as gtag is expected to already be loaded.
@@ -16,35 +17,6 @@ import {LOCAL_STORAGE} from '../../providers/index';
 declare const window: Window & typeof globalThis & {gtag?: Function};
 
 export const STORAGE_KEY = 'docs-accepts-cookies';
-export function setCookieConsent(state: 'denied' | 'granted'): void {
-  try {
-    if (window.gtag) {
-      const consentOptions = {
-        ad_user_data: state,
-        ad_personalization: state,
-        ad_storage: state,
-        analytics_storage: state,
-      };
-
-      if (state === 'denied') {
-        window.gtag('consent', 'default', {
-          ...consentOptions,
-          wait_for_update: 500,
-        });
-      } else if (state === 'granted') {
-        window.gtag('consent', 'update', {
-          ...consentOptions,
-        });
-      }
-    }
-  } catch {
-    if (state === 'denied') {
-      console.error('Unable to set default cookie consent.');
-    } else if (state === 'granted') {
-      console.error('Unable to grant cookie consent.');
-    }
-  }
-}
 
 @Component({
   selector: 'docs-cookie-popup',
